popupmenus: move context package to popupmenus and split and rename interface
* ContextMenu is now PopupMenu
* MenuItem, MenuItemFactory and PopupMenu are now in popupmenus package
* Remaining context package is now subpackage of popupmenus
* DefaultContextMenu is now ContextMenu
