+++ src/test/java/net/kemitix/ldapmanager/popupmenus/context/ContextKeyStrokeHandlerTest.java
@@ -1,4 +1,4 @@
-package net.kemitix.ldapmanager.context;
+package net.kemitix.ldapmanager.popupmenus.context;
 
 import com.googlecode.lanterna.input.KeyStroke;
 import com.googlecode.lanterna.input.KeyType;
